`Hex` with `Hexable` classess added for printing
`Hex` class is a way to printing binary data in easy-to-parse way.
Also Microchip datasheets tells that programmers need to be able to read
and write HEX-files so this library needs it.
Currently defined classes gives only possibility of reading data into
HEX and print it via `Serial`, so the next step is to write HEX to chip.